# perf(depinject): optimize isExportedType by removing allocations and improving performance