optimize isExportedType to reduce allocations and improve performance for basic, composite, and internal types.

- removed all memory allocations in IsExportedType
- improved execution time across all tested types
- significant speedup for internal/unexported type checks

golang 1.25
ubuntu 24.04
```bash
benchstat old.txt new.txt                                                   
goos: linux
goarch: amd64
pkg: cosmossdk.io/depinject
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
                                            │    old.txt    │               new.txt                │
                                            │    sec/op     │    sec/op     vs base                │
IsExportedType/Bool-12                         51.87n ± 18%   51.00n ± 71%        ~ (p=0.912 n=10)
IsExportedType/Uint-12                         43.01n ± 38%   43.64n ±  5%        ~ (p=0.971 n=10)
IsExportedType/Uint8-12                        46.66n ± 46%   46.83n ±  9%        ~ (p=0.971 n=10)
IsExportedType/Uint16-12                       50.49n ± 48%   46.96n ± 13%        ~ (p=0.363 n=10)
IsExportedType/Uint32-12                       50.77n ± 28%   55.35n ±  8%        ~ (p=0.247 n=10)
IsExportedType/Uint64-12                       51.54n ± 11%   48.20n ± 15%        ~ (p=0.280 n=10)
IsExportedType/Int-12                          55.58n ± 14%   38.39n ± 13%  -30.93% (p=0.000 n=10)
IsExportedType/Int8-12                         48.17n ± 16%   45.22n ± 14%        ~ (p=0.190 n=10)
IsExportedType/Int16-12                        49.23n ± 11%   47.84n ± 11%        ~ (p=0.631 n=10)
IsExportedType/Int32-12                        46.97n ± 17%   44.45n ±  4%        ~ (p=0.123 n=10)
IsExportedType/Int64-12                        48.35n ±  9%   48.07n ± 21%        ~ (p=0.631 n=10)
IsExportedType/Float32-12                      53.63n ± 13%   45.45n ± 17%  -15.27% (p=0.043 n=10)
IsExportedType/Float64-12                      57.65n ± 15%   41.68n ± 19%  -27.71% (p=0.000 n=10)
IsExportedType/Complex64-12                    54.83n ± 10%   48.17n ±  8%  -12.15% (p=0.000 n=10)
IsExportedType/Complex128-12                   63.73n ± 16%   54.06n ±  8%  -15.17% (p=0.007 n=10)
IsExportedType/String-12                       48.86n ±  7%   44.69n ± 11%   -8.54% (p=0.035 n=10)
IsExportedType/OSFileMode-12                  200.90n ± 13%   60.45n ±  8%  -69.91% (p=0.000 n=10)
IsExportedType/ArrayOfInt-12                   62.27n ± 16%   66.15n ± 12%        ~ (p=0.066 n=10)
IsExportedType/SliceOfInt-12                   74.54n ± 15%   62.01n ±  7%  -16.82% (p=0.002 n=10)
IsExportedType/ChanOfInt-12                    67.64n ± 14%   62.70n ± 15%        ~ (p=0.123 n=10)
IsExportedType/RecvOnlyChan-12                 70.72n ±  7%   62.95n ±  8%  -10.97% (p=0.007 n=10)
IsExportedType/SendOnlyChan-12                 67.95n ± 10%   60.43n ± 10%  -11.07% (p=0.019 n=10)
IsExportedType/FunctionBasic-12                255.7n ± 13%   250.6n ± 12%        ~ (p=0.838 n=10)
IsExportedType/FunctionVariadic-12             276.7n ±  9%   263.3n ±  5%        ~ (p=0.052 n=10)
IsExportedType/ExportedStruct-12              184.00n ± 10%   67.73n ± 21%  -63.19% (p=0.000 n=10)
IsExportedType/MapStringToExported-12          388.9n ± 28%   142.3n ± 11%  -63.40% (p=0.000 n=10)
IsExportedType/PointerToExported-12           241.80n ± 33%   89.54n ±  4%  -62.97% (p=0.000 n=10)
IsExportedType/Uintptr-12                      72.48n ± 25%   48.59n ±  7%  -32.96% (p=0.000 n=10)
IsExportedType/NilLocationPointer-12           222.8n ± 25%   127.3n ± 31%  -42.88% (p=0.000 n=10)
IsExportedType/UnexportedStruct-12            417.80n ± 16%   59.50n ± 22%  -85.76% (p=0.000 n=10)
IsExportedType/PointerToUnexportedStruct-12   444.60n ± 12%   70.63n ±  9%  -84.11% (p=0.000 n=10)
IsExportedType/InternalTypeGraphviz-12        716.05n ±  9%   74.51n ±  7%  -89.59% (p=0.000 n=10)
IsExportedType/MapWithInternalType-12          730.0n ± 16%   136.4n ± 16%  -81.32% (p=0.000 n=10)
IsExportedType/SliceWithInternalType-12       645.20n ±  8%   97.45n ± 27%  -84.90% (p=0.000 n=10)
geomean                                        106.4n         65.03n        -38.86%

                                            │   old.txt    │                 new.txt                 │
                                            │     B/op     │    B/op     vs base                     │
IsExportedType/Bool-12                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint-12                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint8-12                       0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint16-12                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint32-12                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint64-12                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int-12                         0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int8-12                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int16-12                       0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int32-12                       0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int64-12                       0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Float32-12                     0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Float64-12                     0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Complex64-12                   0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Complex128-12                  0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/String-12                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/OSFileMode-12                  32.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/ArrayOfInt-12                  0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/SliceOfInt-12                  0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/ChanOfInt-12                   0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/RecvOnlyChan-12                0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/SendOnlyChan-12                0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/FunctionBasic-12               0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/FunctionVariadic-12            0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/ExportedStruct-12              32.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/MapStringToExported-12         32.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/PointerToExported-12           32.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/Uintptr-12                     0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/NilLocationPointer-12          32.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/UnexportedStruct-12            64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/PointerToUnexportedStruct-12   64.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/InternalTypeGraphviz-12        144.0 ± 0%       0.0 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/MapWithInternalType-12         144.0 ± 0%       0.0 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/SliceWithInternalType-12       144.0 ± 0%       0.0 ± 0%  -100.00% (p=0.000 n=10)
geomean                                                  ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                                            │   old.txt    │                 new.txt                 │
                                            │  allocs/op   │ allocs/op   vs base                     │
IsExportedType/Bool-12                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint-12                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint8-12                       0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint16-12                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint32-12                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Uint64-12                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int-12                         0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int8-12                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int16-12                       0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int32-12                       0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Int64-12                       0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Float32-12                     0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Float64-12                     0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Complex64-12                   0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/Complex128-12                  0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/String-12                      0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/OSFileMode-12                  1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/ArrayOfInt-12                  0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/SliceOfInt-12                  0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/ChanOfInt-12                   0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/RecvOnlyChan-12                0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/SendOnlyChan-12                0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/FunctionBasic-12               0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/FunctionVariadic-12            0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/ExportedStruct-12              1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/MapStringToExported-12         1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/PointerToExported-12           1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/Uintptr-12                     0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsExportedType/NilLocationPointer-12          1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/UnexportedStruct-12            2.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/PointerToUnexportedStruct-12   2.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/InternalTypeGraphviz-12        3.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/MapWithInternalType-12         3.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
IsExportedType/SliceWithInternalType-12       3.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                                                  ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean
```